Data type semantics: a primitive value is a BigDecimal, and there is a raw value which depends on the type of the value

+++ e/hu.e.compiler/src/hu/e/compiler/internal/model/symbols/ILiteralSymbol.java
@@ -3,6 +3,8 @@
  */
 package hu.e.compiler.internal.model.symbols;
 
+import java.math.BigDecimal;
+
 import hu.e.compiler.ECompilerException;
 
 /**
@@ -11,6 +13,6 @@ import hu.e.compiler.ECompilerException;
  */
 public interface ILiteralSymbol extends ISymbol{
 
-	public long getValue() throws ECompilerException;
+	public BigDecimal getValue() throws ECompilerException;
 	
 }
