+++ e/hu.e.compiler/src/hu/e/compiler/internal/symbols/AbstractSymbolManager.java
@@ -212,7 +212,7 @@ public abstract class AbstractSymbolManager implements ISymbolManager {
 		for(XExpression index : x.getIndex()){
 			ISymbol i = resolve(context, index);
 			if (i.isLiteral()){
-				a = a.getElement(this, (int)((ILiteralSymbol)i).getValue());
+				a = a.getElement(this, (int)((ILiteralSymbol)i).getValue().intValue());
 			}else{
 				throw new ECompilerException(index, "TODO: Only compile-time indexing is supported for now");
 			}
@@ -275,7 +275,7 @@ public abstract class AbstractSymbolManager implements ISymbolManager {
 		
 		if (td instanceof ArrayTypeDef){
 			ArrayTypeDef atd = (ArrayTypeDef)td;
-			int length = (int)((ILiteralSymbol)resolve(context, atd.getSize())).getValue();
+			int length = (int)((ILiteralSymbol)resolve(context, atd.getSize())).getValue().intValue();
 			if (length != symbols.size()) throw new ECompilerException(literalStruct, "Invalid number of elements!");
 			return new ArrayLiteralSymbol(symbols.toArray(new ISymbol[symbols.size()]), td);
 		}
